requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
ImageHash==4.3.0
mwclient==0.10.1
numpy==1.26.4
oauthlib==3.2.2
Pillow==9.2.0
PyWavelets==1.5.0
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.12.0
six==1.16.0
urllib3==2.2.1
